Rename http to httputils to gain clean separation
