+++ httputils/Server.go
@@ -1,4 +1,4 @@
-package http
+package httputils
 
 import "errors"
