+++ httputils/StatusCodes.go
@@ -1,3 +1,3 @@
-package http
+package httputils
 
 const STATUS_CODE_OK = 200
